refactor: move around a bunch of config code, start grouping changes
TODO: Update built-in configs!
TODO: Update docs; add section for config file
