+++ src/canvas/styling.rs
@@ -6,7 +6,10 @@ use tui::style::{Color, Style};
 
 use super::ColourScheme;
 pub use crate::options::Config;
-use crate::{constants::*, options::colours::ConfigColours, utils::error};
+use crate::{
+    options::config::{colours::ColourConfig, palettes::*},
+    utils::error,
+};
 
 pub struct CanvasStyling {
     pub currently_selected_text_colour: Color,
@@ -154,7 +157,7 @@ impl CanvasStyling {
         Ok(canvas_colours)
     }
 
-    pub fn set_colours_from_palette(&mut self, colours: &ConfigColours) -> anyhow::Result<()> {
+    pub fn set_colours_from_palette(&mut self, colours: &ColourConfig) -> anyhow::Result<()> {
         // CPU
         try_set_colour!(self.avg_colour_style, colours, avg_cpu_color);
         try_set_colour!(self.all_colour_style, colours, all_cpu_color);
