+++ src/options/config/battery.rs
@@ -1,0 +1,6 @@
+use serde::Deserialize;
+
+#[derive(Clone, Debug, Default, Deserialize)]
+pub(crate) struct BatteryOptions {
+    pub(crate) battery: Option<bool>,
+}
