+++ src/options/config/process.rs
@@ -4,9 +4,17 @@ use crate::widgets::ProcWidgetColumn;
 
 /// Process column settings.
 #[derive(Clone, Debug, Default, Deserialize)]
-pub struct ProcessConfig {
-    #[serde(default)]
-    pub columns: Vec<ProcWidgetColumn>,
+pub(crate) struct ProcessConfig {
+    pub(crate) case_sensitive: Option<bool>,
+    pub(crate) current_usage: Option<bool>,
+    pub(crate) disable_advanced_kill: Option<bool>,
+    pub(crate) group_processes: Option<bool>,
+    pub(crate) process_command: Option<bool>,
+    pub(crate) regex: Option<bool>,
+    pub(crate) tree: Option<bool>,
+    pub(crate) unnormalized_cpu: Option<bool>,
+    pub(crate) whole_word: Option<bool>,
+    pub(crate) columns: Option<Vec<ProcWidgetColumn>>,
 }
 
 #[cfg(test)]
@@ -18,7 +26,7 @@ mod test {
     fn empty_column_setting() {
         let config = "";
         let generated: ProcessConfig = toml_edit::de::from_str(config).unwrap();
-        assert!(generated.columns.is_empty());
+        assert!(generated.columns.is_none());
     }
 
     #[test]
@@ -29,7 +37,7 @@ mod test {
 
         let generated: ProcessConfig = toml_edit::de::from_str(config).unwrap();
         assert_eq!(
-            generated.columns,
+            generated.columns.unwrap(),
             vec![
                 ProcWidgetColumn::Cpu,
                 ProcWidgetColumn::PidOrCount,
@@ -56,18 +64,22 @@ mod test {
     fn process_column_settings_3() {
         let config = r#"columns = ["Twrite", "T.Write"]"#;
         let generated: ProcessConfig = toml_edit::de::from_str(config).unwrap();
-        assert_eq!(generated.columns, vec![ProcWidgetColumn::TotalWrite; 2]);
+        let columns = generated.columns.unwrap();
+        assert_eq!(columns, vec![ProcWidgetColumn::TotalWrite; 2]);
 
         let config = r#"columns = ["Tread", "T.read"]"#;
         let generated: ProcessConfig = toml_edit::de::from_str(config).unwrap();
-        assert_eq!(generated.columns, vec![ProcWidgetColumn::TotalRead; 2]);
+        let columns = generated.columns.unwrap();
+        assert_eq!(columns, vec![ProcWidgetColumn::TotalRead; 2]);
 
         let config = r#"columns = ["read", "rps", "r/s"]"#;
         let generated: ProcessConfig = toml_edit::de::from_str(config).unwrap();
-        assert_eq!(generated.columns, vec![ProcWidgetColumn::ReadPerSecond; 3]);
+        let columns = generated.columns.unwrap();
+        assert_eq!(columns, vec![ProcWidgetColumn::ReadPerSecond; 3]);
 
         let config = r#"columns = ["write", "wps", "w/s"]"#;
         let generated: ProcessConfig = toml_edit::de::from_str(config).unwrap();
-        assert_eq!(generated.columns, vec![ProcWidgetColumn::WritePerSecond; 3]);
+        let columns = generated.columns.unwrap();
+        assert_eq!(columns, vec![ProcWidgetColumn::WritePerSecond; 3]);
     }
 }
